chore: Improve message for capitalizing pronoun I
Old message was "First-person singular pronouns must be capitalized."

But English has two first-person singular pronouns: I & me.

Only I is capitalized. I changed the message to "The first-person singular subject pronoun must be capitalized."

"I" is also called the nomininative. I chose "subject" because it's a little less jargony for people less familiar with grammar and linguistics, and it's shorter when the message is already long.