feat: add `just sample` to list all forms of random dictionary words
Modified `harper-cli forms` to accept a space-separated list of annotated words to support this.

`just sample` accepts a numeric argument for how many words to choose from the curated dictionary.

I've already fixed some mistakes with wrong affix annotations in the curated dictionary. This provides a way to check randomly to see if anything looks wrong.

It may actually be more useful for the user dictionary. Let me know if I should add that and suggest names for switches. Another possibility as to also have an option that samples from the combined curated+user dictionaries.

The implementation uses the *nix `jot` command which is included with macOS which does not have `shuf`. Let me know if this solution is broken for Linux etc.