feat: detect "more faster", "most best" (#1507)
* feat: detect "more faster", "most best"

work in progress

* fix: `just format`

* refactor: complete basic featureset

Superlative annotation flag changed from `T` to `^` to match the comparative `>`. Mnemonic comment fields added..

* chore: mark comparatives and superlatives in dictionary

These are the ones whose last letter is doubled since the affix logic doesn't handle that.

Add comments to `affixes.json` linking the annotation flags for the base form comparative and superlative with the deirved term ones.

Add `get_degree` to `WordMetadata` and as a bonus also `get_person` for pronouns.

* Positive degree is not yet queryable as I suspect there could be false positives.

* fix: warning