feat: add letter-case flags to `WordMetadata` (#1578)
* feat: add letter-case flags to `WordMetadata`

(Plus other orthography flags.)

These are all for the "case-folded lexemes" in the dictionary, independent of the spelling in the document. So even at the start of a sentence or in all caps you can match words that are capitalized in the dictionary, etc.

Also added a `harper-cli case-variants` command that will print all the words in the dictionary that fold multiple case variants into a single entry.

* chore: orth. query methods `WordMetadata` and `TokenKind`

* chore: improve docs & case variants cmd, feedback from @86xsk

* fix: adjust `println` usage to appease `Precommit`

* fix: appease Precommit a bit more

* refactor: orthography part of word_metadata.rs