fix: canonicalize workspace_root in configuration (#1713)
* fix: canonicalize workspace_root in configuration

If the LSP client does not provide any mechanism to determine the root
of the workspace, we need to resolve things relative to the current
directory, but previously we tried that as `.` which it turns out
`.try_resolve_in()` does not like.

This changes the behaviour to always `.canonicalize()` the root of the
workspace so that subsequent path resolution starts from an absolute
path.

Fixes #1707

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>

* chore: cargo +1.89.0 clippy --all --fix

Deal with incoming clippy lints.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>

* chore: Deal with mismatched_lifetime_syntaxes lint

This lint is now warn-by-default so deal with it now.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>

---------

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>