Implement full verb form annotations (#1730)
* feat: annotations for verb form properties

* feat: annotations for verb form properties

* feat: verb form ("tense") properties fully propagated

It turns out that the expectatiion that all past participles and present participles can function as adjectives is not true after all: "cared" for instance. This has affected a couple of current tests.

The best fix is unclear. Perhaps separate annotations flags for verbs where the assumptions is true versus false. Perhaps defaulting to true and needing to create the individual form entries separately so they don't get the adjective annotation. It seems this strongly correlates with whether verbs are transitive or intransitive, which we don't yet track. So maybe we should handle both together.

- just "past" for regular verbs
- separate "preterite"/"simple past" and "past participle" for irregular verbs

Also:
- POS tag snapshots now annotate all verb forms
- adjective positive degree marked
- improved error messages for JSON file parse failures
- simplified everyday.rs to use these changes
- new rust lifetime lint annotations

* fix: address the concerns in the code review